Refactor A2A server into submodules
